LRCI-3543 Shift Project Comparator to primarily use a Repository instead of DALO
